+++ packages/backend-core/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @windrun-huaiin/backend-core
 
+## 10.0.1
+
+### Patch Changes
+
+- feat(backend): core service logic sink into common package
+- Updated dependencies
+  - @windrun-huaiin/third-ui@10.1.3
+  - @windrun-huaiin/lib@10.1.0
+
 ## 10.0.0
 
 ### Major Changes